requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
passlib
bcrypt
python-jose
python-multipart